Refactor: split conflict finding per pass
