+++ crates/solidity-v2/outputs/cargo/semantic/src/passes/p1_collect_definitions/mod.rs
@@ -14,7 +14,6 @@ use crate::binder::{
     AssemblyBlock, Binder, Definition, FileScope, ParametersScope, Scope, ScopeId,
 };
 use crate::context::SemanticFile;
-use crate::passes::common::conflicts::find_conflicting_definition;
 
 mod conflicts;
 
@@ -143,7 +142,14 @@ impl<'a, F: SemanticFile> Pass<'a, F> {
     // resolve references to it.
     fn insert_definition_in_scope(&mut self, definition: Definition, scope_id: ScopeId) {
         let symbol = definition.identifier().unparse();
-        if find_conflicting_definition(self.binder, scope_id, symbol, &definition).is_some() {
+        if conflicts::find_conflicting_solidity_definition(
+            self.binder,
+            scope_id,
+            symbol,
+            &definition,
+        )
+        .is_some()
+        {
             self.diagnostics.push(
                 self.current_file.id().to_owned(),
                 definition.identifier().range.clone(),
